ct registry reading bug fixes
git-svn-id: https://subversion.assembla.com/svn/smart-cs/trunk@7397 37136b1e-4cb1-4992-8394-573b48a9bdce

+++ source/extensions/cybertracker/org.wcs.smart.cybertracker/src/org/wcs/smart/cybertracker/util/WinRegistry.java
@@ -15,8 +15,8 @@ import org.wcs.smart.cybertracker.CyberTrackerPlugIn;
 
 @SuppressWarnings("nls")
 public class WinRegistry {
-  public static final int HKEY_CURRENT_USER = 0x80000001;
-  public static final int HKEY_LOCAL_MACHINE = 0x80000002;
+  public static final long HKEY_CURRENT_USER = 0x80000001;
+  public static final long HKEY_LOCAL_MACHINE = 0x80000002;
   public static final int REG_SUCCESS = 0;
   public static final int REG_NOTFOUND = 2;
   public static final int REG_ACCESSDENIED = 5;
@@ -40,39 +40,43 @@ public class WinRegistry {
   static {
     try {
       regOpenKey = userClass.getDeclaredMethod("WindowsRegOpenKey",
-          new Class[] { int.class, byte[].class, int.class });
+          new Class[] { long.class, byte[].class, int.class });
       regOpenKey.setAccessible(true);
+      
       regCloseKey = userClass.getDeclaredMethod("WindowsRegCloseKey",
-          new Class[] { int.class });
+          new Class[] { long.class });
       regCloseKey.setAccessible(true);
+      
       regQueryValueEx = userClass.getDeclaredMethod("WindowsRegQueryValueEx",
-          new Class[] { int.class, byte[].class });
+          new Class[] { long.class, byte[].class });
       regQueryValueEx.setAccessible(true);
+      
       regEnumValue = userClass.getDeclaredMethod("WindowsRegEnumValue",
-          new Class[] { int.class, int.class, int.class });
+          new Class[] { long.class, int.class, int.class });
       regEnumValue.setAccessible(true);
+      
       regQueryInfoKey = userClass.getDeclaredMethod("WindowsRegQueryInfoKey1",
-          new Class[] { int.class });
+          new Class[] { long.class });
       regQueryInfoKey.setAccessible(true);
-      regEnumKeyEx = userClass.getDeclaredMethod(  
-          "WindowsRegEnumKeyEx", new Class[] { int.class, int.class,  
-              int.class });  
+      
+      regEnumKeyEx = userClass.getDeclaredMethod("WindowsRegEnumKeyEx", 
+    		  new Class[] { long.class, int.class, int.class });  
       regEnumKeyEx.setAccessible(true);
-      regCreateKeyEx = userClass.getDeclaredMethod(  
-          "WindowsRegCreateKeyEx", new Class[] { int.class,  
-              byte[].class });  
+      
+      regCreateKeyEx = userClass.getDeclaredMethod("WindowsRegCreateKeyEx", 
+    		  new Class[] { long.class, byte[].class });  
       regCreateKeyEx.setAccessible(true);  
-      regSetValueEx = userClass.getDeclaredMethod(  
-          "WindowsRegSetValueEx", new Class[] { int.class,  
-              byte[].class, byte[].class });  
+      
+      regSetValueEx = userClass.getDeclaredMethod("WindowsRegSetValueEx", 
+    		  new Class[] { long.class, byte[].class, byte[].class });  
       regSetValueEx.setAccessible(true); 
-      regDeleteValue = userClass.getDeclaredMethod(  
-          "WindowsRegDeleteValue", new Class[] { int.class,  
-              byte[].class });  
-      regDeleteValue.setAccessible(true); 
-      regDeleteKey = userClass.getDeclaredMethod(  
-          "WindowsRegDeleteKey", new Class[] { int.class,  
-              byte[].class });  
+      
+      regDeleteValue = userClass.getDeclaredMethod("WindowsRegDeleteValue", 
+    		  new Class[] { long.class, byte[].class });  
+      regDeleteValue.setAccessible(true);
+      
+      regDeleteKey = userClass.getDeclaredMethod("WindowsRegDeleteKey", 
+    		  new Class[] { long.class, byte[].class });  
       regDeleteKey.setAccessible(true); 
     }
     catch (Exception e) {
@@ -92,7 +96,7 @@ public class WinRegistry {
    * @throws IllegalAccessException
    * @throws InvocationTargetException
    */
-  public static String readString(int hkey, String key, String valueName) 
+  public static String readString(long hkey, String key, String valueName) 
     throws IllegalArgumentException, IllegalAccessException,
     InvocationTargetException 
   {
@@ -116,7 +120,7 @@ public class WinRegistry {
    * @throws IllegalAccessException
    * @throws InvocationTargetException
    */
-  public static Map<String, String> readStringValues(int hkey, String key) 
+  public static Map<String, String> readStringValues(long hkey, String key) 
     throws IllegalArgumentException, IllegalAccessException,
     InvocationTargetException 
   {
@@ -140,7 +144,7 @@ public class WinRegistry {
    * @throws IllegalAccessException
    * @throws InvocationTargetException
    */
-  public static List<String> readStringSubKeys(int hkey, String key) 
+  public static List<String> readStringSubKeys(long hkey, String key) 
     throws IllegalArgumentException, IllegalAccessException,
     InvocationTargetException 
   {
@@ -163,18 +167,18 @@ public class WinRegistry {
    * @throws IllegalAccessException
    * @throws InvocationTargetException
    */
-  public static void createKey(int hkey, String key) 
+  public static void createKey(long hkey, String key) 
     throws IllegalArgumentException, IllegalAccessException,
     InvocationTargetException 
   {
-    int [] ret;
+    long[] ret;
     if (hkey == HKEY_LOCAL_MACHINE) {
       ret = createKey(systemRoot, hkey, key);
-      regCloseKey.invoke(systemRoot, new Object[] { Integer.valueOf(ret[0]) });
+      regCloseKey.invoke(systemRoot, new Object[] { Long.valueOf(ret[0]) });
     }
     else if (hkey == HKEY_CURRENT_USER) {
       ret = createKey(userRoot, hkey, key);
-      regCloseKey.invoke(userRoot, new Object[] { Integer.valueOf(ret[0]) });
+      regCloseKey.invoke(userRoot, new Object[] { Long.valueOf(ret[0]) });
     }
     else {
       throw new IllegalArgumentException("hkey=" + hkey);
@@ -195,7 +199,7 @@ public class WinRegistry {
    * @throws InvocationTargetException
    */
   public static void writeStringValue
-    (int hkey, String key, String valueName, String value) 
+    (long hkey, String key, String valueName, String value) 
     throws IllegalArgumentException, IllegalAccessException,
     InvocationTargetException 
   {
@@ -218,7 +222,7 @@ public class WinRegistry {
    * @throws IllegalAccessException
    * @throws InvocationTargetException
    */
-  public static void deleteKey(int hkey, String key) 
+  public static void deleteKey(long hkey, String key) 
     throws IllegalArgumentException, IllegalAccessException,
     InvocationTargetException 
   {
@@ -243,11 +247,11 @@ public class WinRegistry {
    * @throws IllegalAccessException
    * @throws InvocationTargetException
    */
-  public static void deleteValue(int hkey, String key, String value) 
+  public static void deleteValue(long hkey, String key, String value) 
     throws IllegalArgumentException, IllegalAccessException,
     InvocationTargetException 
   {
-    int rc = -1;
+    long rc = -1;
     if (hkey == HKEY_LOCAL_MACHINE) {
       rc = deleteValue(systemRoot, hkey, key, value);
     }
@@ -261,124 +265,124 @@ public class WinRegistry {
 
   // =====================
 
-  private static int deleteValue
-    (Preferences root, int hkey, String key, String value)
+  private static long deleteValue
+    (Preferences root, long hkey, String key, String value)
     throws IllegalArgumentException, IllegalAccessException,
     InvocationTargetException 
   {
-    int[] handles = (int[]) regOpenKey.invoke(root, new Object[] {
-        Integer.valueOf(hkey), toCstr(key), Integer.valueOf(KEY_ALL_ACCESS) });
+    long[] handles = (long[]) regOpenKey.invoke(root, new Object[] {
+    	Long.valueOf(hkey), toCstr(key), Integer.valueOf(KEY_ALL_ACCESS) });
     if (handles[1] != REG_SUCCESS) {
       return handles[1];  // can be REG_NOTFOUND, REG_ACCESSDENIED
     }
     int rc =((Integer) regDeleteValue.invoke(root,  
         new Object[] { 
-          Integer.valueOf(handles[0]), toCstr(value) 
+        		Long.valueOf(handles[0]), toCstr(value) 
           })).intValue();
-    regCloseKey.invoke(root, new Object[] { Integer.valueOf(handles[0]) });
+    regCloseKey.invoke(root, new Object[] { Long.valueOf(handles[0]) });
     return rc;
   }
 
-  private static int deleteKey(Preferences root, int hkey, String key) 
+  private static int deleteKey(Preferences root, long hkey, String key) 
     throws IllegalArgumentException, IllegalAccessException,
     InvocationTargetException 
   {
     int rc =((Integer) regDeleteKey.invoke(root,  
-        new Object[] { Integer.valueOf(hkey), toCstr(key) })).intValue();
+        new Object[] { Long.valueOf(hkey), toCstr(key) })).intValue();
     return rc;  // can REG_NOTFOUND, REG_ACCESSDENIED, REG_SUCCESS
   }
 
-  private static String readString(Preferences root, int hkey, String key, String value)
+  private static String readString(Preferences root, long hkey, String key, String value)
     throws IllegalArgumentException, IllegalAccessException,
     InvocationTargetException 
   {
-    int[] handles = (int[]) regOpenKey.invoke(root, new Object[] {
-        Integer.valueOf(hkey), toCstr(key), Integer.valueOf(KEY_READ) });
+    long[] handles = (long[]) regOpenKey.invoke(root, new Object[] {
+        Long.valueOf(hkey), toCstr(key), Integer.valueOf(KEY_READ) });
     if (handles[1] != REG_SUCCESS) {
       return null; 
     }
     byte[] valb = (byte[]) regQueryValueEx.invoke(root, new Object[] {
-        Integer.valueOf(handles[0]), toCstr(value) });
-    regCloseKey.invoke(root, new Object[] { Integer.valueOf(handles[0]) });
+        Long.valueOf(handles[0]), toCstr(value) });
+    regCloseKey.invoke(root, new Object[] { Long.valueOf(handles[0]) });
     return (valb != null ? new String(valb).trim() : null);
   }
 
   private static Map<String,String> readStringValues
-    (Preferences root, int hkey, String key)
+    (Preferences root, long hkey, String key)
     throws IllegalArgumentException, IllegalAccessException,
     InvocationTargetException 
   {
     HashMap<String, String> results = new HashMap<String,String>();
-    int[] handles = (int[]) regOpenKey.invoke(root, new Object[] {
-        Integer.valueOf(hkey), toCstr(key), Integer.valueOf(KEY_READ) });
+    long[] handles = (long[]) regOpenKey.invoke(root, new Object[] {
+    		Long.valueOf(hkey), toCstr(key), Integer.valueOf(KEY_READ) });
     if (handles[1] != REG_SUCCESS) {
       return null;
     }
-    int[] info = (int[]) regQueryInfoKey.invoke(root,
-        new Object[] { Integer.valueOf(handles[0]) });
+    long[] info = (long[]) regQueryInfoKey.invoke(root,
+        new Object[] { Long.valueOf(handles[0]) });
 
-    int count = info[0]; // count  
-    int maxlen = info[3]; // value length max
+    long count = info[0]; // count  
+    long maxlen = info[3]; // value length max
     for(int index=0; index<count; index++)  {
       byte[] name = (byte[]) regEnumValue.invoke(root, new Object[] {
-          Integer.valueOf
-            (handles[0]), Integer.valueOf(index), Integer.valueOf(maxlen + 1)});
+    		  Long.valueOf
+            (handles[0]), Integer.valueOf(index), Long.valueOf(maxlen + 1)});
       String value = readString(hkey, key, new String(name));
       results.put(new String(name).trim(), value);
     }
-    regCloseKey.invoke(root, new Object[] { Integer.valueOf(handles[0]) });
+    regCloseKey.invoke(root, new Object[] { Long.valueOf(handles[0]) });
     return results;
   }
 
   private static List<String> readStringSubKeys
-    (Preferences root, int hkey, String key)
+    (Preferences root, long hkey, String key)
     throws IllegalArgumentException, IllegalAccessException,
     InvocationTargetException 
   {
     List<String> results = new ArrayList<String>();
-    int[] handles = (int[]) regOpenKey.invoke(root, new Object[] {
-        Integer.valueOf(hkey), toCstr(key), Integer.valueOf(KEY_READ) 
+    long[] handles = (long[]) regOpenKey.invoke(root, new Object[] {
+    		Long.valueOf(hkey), toCstr(key), Integer.valueOf(KEY_READ) 
         });
     if (handles[1] != REG_SUCCESS) {
       return null;
     }
-    int[] info = (int[]) regQueryInfoKey.invoke(root,
-        new Object[] { Integer.valueOf(handles[0]) });
+    long[] info = (long[]) regQueryInfoKey.invoke(root,
+        new Object[] { Long.valueOf(handles[0]) });
 
-    int count  = info[0]; // Fix: info[2] was being used here with wrong results. Suggested by davenpcj, confirmed by Petrucio
-    int maxlen = info[3]; // value length max
+    long count  = info[0]; // Fix: info[2] was being used here with wrong results. Suggested by davenpcj, confirmed by Petrucio
+    long maxlen = info[3]; // value length max
     for(int index=0; index<count; index++)  {
       byte[] name = (byte[]) regEnumKeyEx.invoke(root, new Object[] {
-          Integer.valueOf
-            (handles[0]), Integer.valueOf(index), Integer.valueOf(maxlen + 1)
+    		  Long.valueOf
+            (handles[0]), Integer.valueOf(index), Long.valueOf(maxlen + 1)
           });
       results.add(new String(name).trim());
     }
-    regCloseKey.invoke(root, new Object[] { Integer.valueOf(handles[0]) });
+    regCloseKey.invoke(root, new Object[] { Long.valueOf(handles[0]) });
     return results;
   }
 
-  private static int [] createKey(Preferences root, int hkey, String key)
+  private static long [] createKey(Preferences root, long hkey, String key)
     throws IllegalArgumentException, IllegalAccessException,
     InvocationTargetException 
   {
-    return  (int[]) regCreateKeyEx.invoke(root,
-        new Object[] { Integer.valueOf(hkey), toCstr(key) });
+    return  (long[]) regCreateKeyEx.invoke(root,
+        new Object[] { Long.valueOf(hkey), toCstr(key) });
   }
 
   private static void writeStringValue 
-    (Preferences root, int hkey, String key, String valueName, String value) 
+    (Preferences root, long hkey, String key, String valueName, String value) 
     throws IllegalArgumentException, IllegalAccessException,
     InvocationTargetException 
   {
-    int[] handles = (int[]) regOpenKey.invoke(root, new Object[] {
-        Integer.valueOf(hkey), toCstr(key), Integer.valueOf(KEY_ALL_ACCESS) });
+	  long[] handles = (long[]) regOpenKey.invoke(root, new Object[] {
+			  Long.valueOf(hkey), toCstr(key), Integer.valueOf(KEY_ALL_ACCESS) });
 
     regSetValueEx.invoke(root,  
         new Object[] { 
-          Integer.valueOf(handles[0]), toCstr(valueName), toCstr(value) 
+          Long.valueOf(handles[0]), toCstr(valueName), toCstr(value) 
           }); 
-    regCloseKey.invoke(root, new Object[] { Integer.valueOf(handles[0]) });
+    regCloseKey.invoke(root, new Object[] { Long.valueOf(handles[0]) });
   }
 
   // utility
